Coding Platform with AI Hints/SQL Coding Platform with AI Hints/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
databases
aiosqlite 
aiomysql
python-dotenv
nh3
beautifulsoup4
langchain
langchain-openai
openai
passlib
bcrypt==4.0.1
python-jose
python-multipart
pydantic[email]
langchain-google-genai
cssutils